% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoms_makedf.R
\name{anoms_makedf}
\alias{anoms_makedf}
\title{Make deficit and surplus data frames}
\usage{
anoms_makedf(file_path, pix_size, coun_area, n_coun, min_pop)
}
\arguments{
\item{file_path}{A file path that points to a .csv file}

\item{pix_size}{Minimum number of country pixels}

\item{coun_area}{Country land area in square kilometers}

\item{n_coun}{Number of countries that will display on a plot}

\item{min_pop}{Country population}
}
\value{
A water deficit data frame and water surplus data frame
}
\description{
Make deficit and surplus data frames
}
\examples{
file_path <- "~/population_water_anomaly_summary_201801.csv"
pix_size <- 10
n_coun <- 10
}
